+++ ps1-gui/src/emuthread/audio.rs
@@ -16,7 +16,7 @@ pub fn new_spec(config: &AudioConfig) -> AudioSpecDesired {
     }
 }
 
-pub type AudioQueue = Arc<Mutex<VecDeque<(f32, f32)>>>;
+pub type AudioQueue = Arc<Mutex<VecDeque<(i16, i16)>>>;
 
 pub struct QueueAudioCallback {
     audio_queue: AudioQueue,
@@ -29,12 +29,12 @@ impl QueueAudioCallback {
 }
 
 impl AudioCallback for QueueAudioCallback {
-    type Channel = f32;
+    type Channel = i16;
 
     fn callback(&mut self, out: &mut [Self::Channel]) {
         let mut queue = self.audio_queue.lock().unwrap();
         for chunk in out.chunks_exact_mut(2) {
-            let (l, r) = queue.pop_front().unwrap_or((0.0, 0.0));
+            let (l, r) = queue.pop_front().unwrap_or((0, 0));
             chunk[0] = l;
             chunk[1] = r;
         }
@@ -65,10 +65,10 @@ impl QueueAudioOutput {
 impl AudioOutput for QueueAudioOutput {
     type Err = Never;
 
-    fn queue_samples(&mut self, samples: &[(f64, f64)]) -> Result<(), Self::Err> {
+    fn queue_samples(&mut self, samples: &[(i16, i16)]) -> Result<(), Self::Err> {
         let mut queue = self.audio_queue.lock().unwrap();
         for &(sample_l, sample_r) in samples {
-            queue.push_back((sample_l as f32, sample_r as f32));
+            queue.push_back((sample_l, sample_r));
         }
 
         Ok(())
